refactor(mobile): extract refresh-on-401 middleware + split graphql by domain (Phase 9)
Phase 9 of the mobile refactor plan.

Auth client improvements:
- Extract refresh-on-401 middleware with DI (setRefreshHandler) in
  lib/graphql/middleware/refresh-on-401.ts. Breaks the circular dependency
  between client.ts and auth-store — the middleware now accepts a refresh
  callback that auth-store wires during initialize.
- client.ts drops inline 401 retry logic and delegates to the middleware.
- Concurrent 401s dedupe into a single refresh call (locked by test).

Token migration:
- Export migrateLegacyTokens from lib/auth/secure-storage so callers can
  trigger it explicitly. Remove the inline call from getToken() so the
  migration only runs once per launch, before the first token read.
- auth-store.initialize now runs migrateLegacyTokens BEFORE getToken(),
  ordering locked by a test in stores/auth-store.test.ts.

GraphQL domain split:
- Split lib/graphql/mutations.ts (22 exports) into five domain files:
  auth, me, dog, encounter, walk under lib/graphql/mutations/.
- Split lib/graphql/queries.ts (7 exports) into five domain files:
  me, dog, walk, friendship, encounter under lib/graphql/queries/.
- Rewrite 15 importers (hooks, stores, components) to use the new paths.

317/317 jest tests pass.
