+++ apps/mobile/lib/graphql/queries/me.ts
@@ -1,0 +1,23 @@
+import { gql } from 'graphql-request';
+
+export const ME_QUERY = gql`
+  query Me {
+    me {
+      id
+      displayName
+      avatarUrl
+      encounterDetectionEnabled
+      dogs {
+        id
+        name
+        breed
+        gender
+        birthDate { year month day }
+        photoUrl
+        role
+        createdAt
+      }
+      createdAt
+    }
+  }
+`;
